#318: update tests/command_download_*.py
