+++ tests/command_download_kattis.py
@@ -12,10 +12,12 @@ class DownloadKattisTest(unittest.TestCase):
         self.snippet_call_download(
             'https://open.kattis.com/contests/asiasg15prelwarmup/problems/8queens', [
                 {
+                    "name": "example00",
                     "input": "*.......\n..*.....\n....*...\n......*.\n.*......\n.......*\n.....*..\n...*....\n",
                     "output": "invalid\n"
                 },
                 {
+                    "name": "example01",
                     "input": "*.......\n......*.\n....*...\n.......*\n.*......\n...*....\n.....*..\n..*.....\n",
                     "output": "valid\n"
                 },
@@ -25,6 +27,7 @@ class DownloadKattisTest(unittest.TestCase):
         self.snippet_call_download(
             'https://hanoi18.kattis.com/problems/amazingadventures', [
                 {
+                    "name": "1",
                     "input": "3 3\n1 1\n3 3\n2 1\n2 2\n\n3 4\n1 1\n3 4\n2 1\n1 2\n\n2 2\n2 1\n2 2\n1 2\n1 1\n\n0 0\n",
                     "output": "YES\nRRUULLD\nNO\nYES\nRD\n"
                 },
